Multiple bug fixes and UX improvements
-Creating a routine now gives the option of "basic" or "advanced", where the latter lets you build orchestrations
-Improved sized of titles on mobile
-Fixed help button menus from overflowing on mobile
-

+++ packages/server/src/models/node.ts
@@ -58,7 +58,7 @@ export const nodeVerifier = () => ({
      * Also doubles as a way to verify that all nodes are in the same routine.
      * @returns Promise with routineId, so we can use it later
      */
-    async authorizedCheck(userId: string, nodeIds: string[], prisma: PrismaType): Promise<string> {
+    async authorizedCheck(userId: string, nodeIds: string[], prisma: PrismaType): Promise<string | null> {
         let nodes = await prisma.node.findMany({
             where: {
                 AND: [
@@ -86,7 +86,18 @@ export const nodeVerifier = () => ({
         });
         // Check that all nodes are in the same routine
         const uniqueRoutineIds = new Set(nodes.map(node => node.routineId))
-        if (uniqueRoutineIds.size === 0) throw new CustomError(CODE.Unauthorized, 'You do not own all nodes');
+        // If not routine ids found, check if all nodes are new
+        if (uniqueRoutineIds.size === 0) {
+            const newNodes = await prisma.node.count({
+                where: {
+                    id: { in: nodeIds },
+                }
+            })
+            // If all nodes are new, return null for routineId
+            if (newNodes === 0) return null;
+            // If not all nodes are new, then there must be one that is not new and not owned by the user
+            else throw new CustomError(CODE.Unauthorized, 'You do not own all nodes');
+        }
         if (uniqueRoutineIds.size > 1) throw new CustomError(CODE.InvalidArgs, 'All nodes must be in the same routine!');
         return uniqueRoutineIds.values().next().value;
     },
@@ -112,6 +123,7 @@ export const nodeVerifier = () => ({
 export const nodeMutater = (prisma: PrismaType, verifier: any) => ({
     async toDBShape(userId: string | null, data: NodeCreateInput | NodeUpdateInput): Promise<any> {
         let nodeData: { [x: string]: any } = {
+            id: data.id,
             columnIndex: data.columnIndex,
             routineId: data.routineId,
             rowIndex: data.rowIndex,
@@ -357,6 +369,7 @@ export const nodeMutater = (prisma: PrismaType, verifier: any) => ({
                 let routineRel = await routineModel.relationshipBuilder(userId, data, isAdd, 'routine')
                 if (routineRel?.connect) routineRel.connect = routineRel.connect[0];
                 result.push({
+                    id: data.id ?? undefined,
                     index: data.index,
                     isOptional: data.isOptional,
                     routine: routineRel,
@@ -438,12 +451,12 @@ export const nodeMutater = (prisma: PrismaType, verifier: any) => ({
         const deleteNodeIds = deleteMany ?? [];
         const allNodeIds = [...createNodeIds, ...updateNodeIds, ...deleteNodeIds];
         if (allNodeIds.length === 0) return;
-        const routineId = await verifier.authorizedCheck(userId, allNodeIds, prisma);
+        const routineId: string | null = await verifier.authorizedCheck(userId, allNodeIds, prisma);
         if (createMany) {
             createMany.forEach(input => nodeCreate.validateSync(input, { abortEarly: false }));
             createMany.forEach(input => verifier.profanityCheck(input));
             // Check if will pass max nodes (on routine) limit
-            await verifier.maximumCheck(routineId, (createMany?.length ?? 0) - (deleteMany?.length ?? 0), prisma);
+            if (routineId) await verifier.maximumCheck(routineId, (createMany?.length ?? 0) - (deleteMany?.length ?? 0), prisma);
         }
         if (updateMany) {
             updateMany.forEach(input => nodeUpdate.validateSync(input.data, { abortEarly: false }));
